tests/wpt/tests/css/css-fonts/parsing/font-width-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing font-width with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#font-width-prop">
<meta name="assert" content="font-width supports only the grammar 'normal | <percentage [0,∞]> | ultra-condensed | extra-condensed | condensed | semi-condensed | semi-expanded | expanded | extra-expanded | ultra-expanded'.">
<meta name="assert" content="Values less than 0% are invalid.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
for (const property of ['font-width','font-stretch']) {
    test_invalid_value(property, 'auto');
    test_invalid_value(property, 'normal, ultra-condensed');
    test_invalid_value(property, 'condensed expanded');
    test_invalid_value(property, '-50%');
}
</script>
</body>
</html>
